+++ front/stat.tracking.php
@@ -33,6 +33,7 @@
  * ---------------------------------------------------------------------
  */
 
+use Glpi\Application\View\TemplateRenderer;
 use Glpi\Stat\Data\Location\StatDataClosed;
 use Glpi\Stat\Data\Location\StatDataLate;
 use Glpi\Stat\Data\Location\StatDataOpened;
@@ -49,21 +50,8 @@ if (!$item = getItemForItemtype($_GET['itemtype'])) {
     exit;
 }
 
-if (empty($_GET["type"])) {
-    $_GET["type"] = "user";
-}
-
-if (empty($_GET["showgraph"])) {
-    $_GET["showgraph"] = 0;
-}
-
-if (empty($_GET["value2"])) {
-    $_GET["value2"] = 0;
-}
-
 if (empty($_GET["date1"]) && empty($_GET["date2"])) {
-    $year              = date("Y") - 1;
-    $_GET["date1"] = date("Y-m-d", mktime(1, 0, 0, date("m"), date("d"), $year));
+    $_GET["date1"] = date("Y-m-d", mktime(1, 0, 0, date("m"), date("d"), date("Y") - 1));
     $_GET["date2"] = date("Y-m-d");
 }
 
@@ -77,22 +65,18 @@ if (
     $_GET["date2"] = $tmp;
 }
 
-if (!isset($_GET["start"])) {
-    $_GET["start"] = 0;
-}
-
-$stat = new Stat();
-Stat::title();
-
-$requester = ['user'               => ['title' => _n('Requester', 'Requesters', 1)],
-    'users_id_recipient' => ['title' => __('Writer')],
-    'group'              => ['title' => Group::getTypeName(1)],
-    'group_tree'         => ['title' => __('Group tree')],
-    'usertitles_id'      => ['title' => _x('person', 'Title')],
-    'usercategories_id'  => ['title' => _n('Category', 'Categories', 1)]
+$params = [
+    'itemtype'  => $_GET["itemtype"] ?? "",
+    'type'      => $_GET["type"] ?? "user",
+    'date1'     => $_GET["date1"],
+    'date2'     => $_GET["date2"],
+    'value2'    => $_GET["value2"] ?? 0,
+    'start'     => $_GET["start"] ?? 0,
+    'showgraph' => $_GET["showgraph"] ?? 0,
 ];
 
-$caract    = ['itilcategories_id'   => ['title' => _n('Category', 'Categories', 1)],
+$caract = [
+    'itilcategories_id'   => ['title' => _n('Category', 'Categories', 1)],
     'itilcategories_tree' => ['title' => __('Category tree')],
     'urgency'             => ['title' => __('Urgency')],
     'impact'              => ['title' => __('Impact')],
@@ -100,26 +84,29 @@ $caract    = ['itilcategories_id'   => ['title' => _n('Category', 'Categories', 
     'solutiontypes_id'    => ['title' => SolutionType::getTypeName(1)]
 ];
 
-if ($_GET['itemtype'] == 'Ticket') {
+if ($params['itemtype'] == 'Ticket') {
     $caract['type']            = ['title' => _n('Type', 'Types', 1)];
     $caract['requesttypes_id'] = ['title' => RequestType::getTypeName(1)];
     $caract['locations_id']    = ['title' => Location::getTypeName(1)];
     $caract['locations_tree']  = ['title' => __('Location tree')];
 }
 
-
-$items = [_n('Requester', 'Requesters', 1)       => $requester,
+$items = [
+    _n('Requester', 'Requesters', 1) => [
+        'user'               => ['title' => _n('Requester', 'Requesters', 1)],
+        'users_id_recipient' => ['title' => __('Writer')],
+        'group'              => ['title' => Group::getTypeName(1)],
+        'group_tree'         => ['title' => __('Group tree')],
+        'usertitles_id'      => ['title' => _x('person', 'Title')],
+        'usercategories_id'  => ['title' => _n('Category', 'Categories', 1)]
+    ],
     __('Characteristics') => $caract,
-    __('Assigned to')     => ['technicien'
-                                                   => ['title' => __('Technician as assigned')],
-        'technicien_followup'
-                                                   => ['title' => __('Technician in tasks')],
-        'groups_id_assign'
-                                                   => ['title' => Group::getTypeName(1)],
-        'groups_tree_assign'
-                                                   => ['title' => __('Group tree')],
-        'suppliers_id_assign'
-                                                   => ['title' => Supplier::getTypeName(1)]
+    __('Assigned to') => [
+        'technicien'          => ['title' => __('Technician as assigned')],
+        'technicien_followup' => ['title' => __('Technician in tasks')],
+        'groups_id_assign'    => ['title' => Group::getTypeName(1)],
+        'groups_tree_assign'  => ['title' => __('Group tree')],
+        'suppliers_id_assign' => ['title' => Supplier::getTypeName(1)]
     ]
 ];
 
@@ -130,88 +117,56 @@ foreach ($items as $label => $tab) {
     }
 }
 
-echo "<div class='center'><form method='get' name='form' action='stat.tracking.php'>";
-// Keep it first param
-echo "<input type='hidden' name='itemtype' value=\"" . $_GET["itemtype"] . "\">";
-
-echo "<table class='tab_cadre_fixe'>";
-echo "<tr class='tab_bg_2'><td rowspan='2' class='center' width='30%'>";
-Dropdown::showFromArray('type', $values, ['value' => $_GET['type']]);
-echo "</td>";
-echo "<td class='right'>" . __('Start date') . "</td><td>";
-Html::showDateField("date1", ['value' => $_GET["date1"]]);
-echo "</td>";
-echo "<td class='right'>" . __('Show graphics') . "</td>";
-echo "<td rowspan='2' class='center'>";
-echo "<input type='submit' class='btn btn-primary' name='submit' value=\"" . __s('Display report') . "\"></td>" .
-     "</tr>";
-
-echo "<tr class='tab_bg_2'><td class='right'>" . __('End date') . "</td><td>";
-Html::showDateField("date2", ['value' => $_GET["date2"]]);
-echo "</td><td class='center'>";
-echo "<input type='hidden' name='value2' value='" . $_GET["value2"] . "'>";
-Dropdown::showYesNo('showgraph', $_GET['showgraph']);
-echo "</td></tr>";
-echo "</table>";
-// form using GET method : CRSF not needed
-echo "</form>";
-echo "</div>";
-
-$val    = Stat::getItems(
-    $_GET["itemtype"],
-    $_GET["date1"],
-    $_GET["date2"],
-    $_GET["type"],
-    $_GET["value2"]
-);
-$params = ['type'   => $_GET["type"],
-    'date1'  => $_GET["date1"],
-    'date2'  => $_GET["date2"],
-    'value2' => $_GET["value2"],
-    'start'  => $_GET["start"]
-];
-
+Stat::title();
 Html::printPager(
-    $_GET['start'],
+    $params['start'],
     count($val),
     $CFG_GLPI['root_doc'] . '/front/stat.tracking.php',
-    "date1=" . $_GET["date1"] . "&amp;date2=" . $_GET["date2"] . "&amp;type=" . $_GET["type"] .
-                    "&amp;showgraph=" . $_GET["showgraph"] . "&amp;itemtype=" . $_GET["itemtype"] .
-                    "&amp;value2=" . $_GET['value2'],
+    http_build_query($params),
     'Stat',
     $params
 );
 
-if (!$_GET['showgraph']) {
+TemplateRenderer::getInstance()->display('pages/assistance/stats/tracking_form.html.twig', [
+    'itemtype'  => $params["itemtype"],
+    'types'     => $values,
+    'type'      => $params['type'],
+    'date1'     => $params["date1"],
+    'date2'     => $params["date2"],
+    'value2'    => $params["value2"],
+    'showgraph' => $params["showgraph"],
+]);
+
+$val = Stat::getItems(
+    $params["itemtype"],
+    $params["date1"],
+    $params["date2"],
+    $params["type"],
+    $params["value2"]
+);
+
+if (!$params['showgraph']) {
     Stat::showTable(
-        $_GET["itemtype"],
-        $_GET["type"],
-        $_GET["date1"],
-        $_GET["date2"],
-        $_GET['start'],
+        $params["itemtype"],
+        $params["type"],
+        $params["date1"],
+        $params["date2"],
+        $params['start'],
         $val,
-        $_GET['value2']
+        $params['value2']
     );
 } else {
-    $data = Stat::getData(
-        $_GET["itemtype"],
-        $_GET["type"],
-        $_GET["date1"],
-        $_GET["date2"],
-        $_GET['start'],
-        $val,
-        $_GET['value2']
-    );
-
     $data_params = [
-        'itemtype' => $_GET['itemtype'],
-        'type'     => $_GET["type"],
-        'date1'    => $_GET['date1'],
-        'date2'    => $_GET['date2'],
-        'start'    => $_GET['start'],
+        'itemtype' => $params['itemtype'],
+        'type'     => $params["type"],
+        'date1'    => $params['date1'],
+        'date2'    => $params['date2'],
+        'start'    => $params['start'],
         'val'      => $val,
-        'value2'   => $_GET['value2'],
+        'value2'   => $params['value2'],
     ];
+
+    $stat = new Stat();
 
     $stat->displayPieGraphFromData(new StatDataOpened($data_params));
     $stat->displayPieGraphFromData(new StatDataSolved($data_params));